feat: Add my solution to 06-flex-layout
This seems to work, but there is inconsistent behavior when shrinking
the page: In dev tools, the image collapses as expected and the footer
stays in place. With the device toolbar, it starts to shrink the whole
page and the footer rises.

Upon review, the solution CSS breaks the same way, so I am leaving it.